Import of SAMPO/zxid-1.01 from CPAN.
gitpan-cpan-distribution: zxid
gitpan-cpan-version:      1.01
gitpan-cpan-path:         SAMPO/zxid-1.01.tgz
gitpan-cpan-author:       SAMPO
gitpan-cpan-maturity:     released

+++ zxdecode.c
@@ -15,11 +15,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <errno.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <signal.h>
-#include <fcntl.h>
 
+#include "platform.h"
 #include "errmac.h"
 #include "zx.h"
 #include "zxid.h"
@@ -454,7 +451,7 @@ b64_dec:
     /* msglen = q - msg; */
     p = unbase64_raw(msg, q, msg, zx_std_index_64);  /* inplace */
     *p = 0;
-    D("Unbase64 Msg(%s) x=%x len=%d (n.b. message data may be binary at this point)", msg, *msg, p-msg);
+    D("Unbase64 Msg(%s) x=%x len=%d (n.b. message data may be binary at this point)", msg, *msg, ((int)(p-msg)));
     break;
   case 2:
     if (*msg == '<') {
@@ -485,7 +482,7 @@ decompress:
     for (p2 = p-1; m2 < p2; --p2)
       if (!ONE_OF_4(*p2, ' ', '\t', '\015', '\012'))
 	break;
-    D("Msg_minus_whitespace(%.*s) start=%x end=%x", p2-m2+1, m2, *m2, *p2);
+    D("Msg_minus_whitespace(%.*s) start=%x end=%x", ((int)(p2-m2+1)), m2, *m2, *p2);
     
     if (*m2 == '<' && *p2 == '>') {  /* POST profiles do not compress the payload */
       len = p2 - m2 + 1;
