+++ zxidcdc.c
@@ -1,5 +1,5 @@
 /* zxidcdc.c  -  Handwritten functions for Common Domain Cookie handling at SP
- * Copyright (c) 2010 Sampo Kellomaki (sampo@iki.fi), All Rights Reserved.
+ * Copyright (c) 2010-2011 Sampo Kellomaki (sampo@iki.fi), All Rights Reserved.
  * Copyright (c) 2006-2008 Symlabs (symlabs@symlabs.com), All Rights Reserved.
  * Author: Sampo Kellomaki (sampo@iki.fi)
  * This is confidential unpublished proprietary source code of the author.
@@ -16,7 +16,7 @@
 #include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
-
+#include "platform.h"
 #include "errmac.h"
 #include "zxid.h"
 #include "zxidutil.h"
@@ -112,7 +112,7 @@ int zxid_cdc_check(zxid_conf* cf, zxid_cgi* cgi)
     len = p ? p-q : strlen(q);
     
     if (SIMPLE_BASE64_PESSIMISTIC_DECODE_LEN(len) > sizeof(eid)-1) {
-      ERR("EntityID len=%d larger than built in limit=%d. Base64 len=%d", SIMPLE_BASE64_PESSIMISTIC_DECODE_LEN(len), sizeof(eid)-1, len);
+      ERR("EntityID len=%d larger than built in limit=%d. Base64 len=%d", SIMPLE_BASE64_PESSIMISTIC_DECODE_LEN(len), (int)sizeof(eid)-1, len);
       continue;
     }
     q = unbase64_raw(q, q + len, eid, zx_std_index_64);
